feat(nvim and zshrc): update nvim plugins and add docker alias
This commit does two main things. First, it updates the commit hashes
for the Neovim plugins 'lazy.nvim', 'nvim-lspconfig', 'nvim-treesitter',
and 'render-markdown'. This ensures we're using the latest versions of
these plugins.

Second, it adds a new alias for Docker in the .zshrc file. The alias 'dcb'
stands for 'docker compose build', which will be a handy shortcut for
building Docker images.